+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - feat: `fma`, `isfinite`, `isinf`, and `isnan` math function added by @NAThompson
 - feat: `quantity_point` support added for `quantity_cast` and `value_cast`
+- feat: `value_cast<Unit, Representation>` added
 - (!) refactor: `zero_Fahrenheit` renamed to `zeroth_Fahrenheit`
 - refactor: math functions constraints refactored
 - fix: `QuantityLike` conversions required `Q::rep` instead of using one provided by `quantity_like_traits`